+++ src/test/java/vehiculos/Camion.java
@@ -4,7 +4,7 @@ public class Camion extends Vehiculo{
 	private int ejes;
 	private static int cantidadCamiones;
 	
-	public Camion (String placa, String nombre, long precio, int peso,Fabricante fabricante,int ejes){
+	public Camion (String placa, String nombre,int precio, int peso,Fabricante fabricante,int ejes){
 		super(placa,2,80,nombre,precio,peso,"4X2",fabricante);
 		this.setEjes(ejes);
 		Camion.cantidadCamiones++;
