+++ src/test/java/vehiculos/Camioneta.java
@@ -4,7 +4,7 @@ public class Camioneta extends Vehiculo{
 	private boolean volco;
 	private static int cantidadCamionetas;
 	
-	public Camioneta (String placa,int puertas, String nombre, long precio, int peso,Fabricante fabricante, boolean volco){
+	public Camioneta (String placa,int puertas, String nombre, int precio, int peso,Fabricante fabricante, boolean volco){
 		super(placa,puertas,90,nombre,precio,peso,"4X4",fabricante);
 		this.setVolco(volco);
 		Camioneta.cantidadCamionetas++;
